feat: runtime instruction detection

+++ src/smith_waterman/simd/indices.rs
@@ -115,7 +115,7 @@ mod tests {
 
     fn run_single_indices(needle: &str, haystack: &str) -> Vec<usize> {
         let haystacks = [haystack; 1];
-        let (_, score_matrices, _) = smith_waterman::<u8, 16, 1>(needle, &haystacks);
+        let (_, score_matrices, _) = smith_waterman::<u16, 16, 1>(needle, &haystacks);
         let indices = char_indices_from_scores(&score_matrices);
         indices[0].clone()
     }
@@ -142,7 +142,7 @@ mod tests {
             "toolbar",
         ];
 
-        let (_, score_matrices, _) = smith_waterman::<u8, 16, 16>(needle, &haystacks);
+        let (_, score_matrices, _) = smith_waterman::<u16, 16, 16>(needle, &haystacks);
         let indices = char_indices_from_scores(&score_matrices);
         for indices in indices.into_iter() {
             assert_eq!(indices, [0])
